[ style ] 버튼 옮기기..
